ISSUE-47176175: Branch commit.Migrating subscription repository (Before)

+++ tasks_branches/branched_from_v1_1/ISSUE-47176175_Refactor_DAL/Code/Main/ermeX.Domain/Component/IRegisterComponents.cs
@@ -13,6 +13,6 @@ namespace ermeX.Domain.Component
 		/// <returns>a value indicating whether the component is new or not</returns>
         bool CreateRemoteComponent(Guid remoteComponentId, string ip, int port);
        
-        void CreateLocalComponent(int port);
+        void CreateLocalComponent(ushort port);
     }
 }
